Change the order of ifdef and define
When you take the actual example file, and uncomment the line #define OUTPUT_READABLE_YAWPITCHROLL and try to compile this code, you get an error "ypr was not declared in this scope". This is caused by a wrong order of #define and #ifdef.
With this patch, it is possible to compile in the correct way.