Define BUFFER_LENGTH if not already present
Default buffer size for I2C transmissions is 32 if not otherwise set.
This is a band-aid fix since there is almost certainly a more efficient
way to handle this, possibly by removing the chunking process entirely
from the I2Cdev layer if Wire implements it now based on its own
internal buffering (not sure).