Added PID Calibration routine with 2016 code
This allows us to compare the optimized PID code with the traditional averaging method original 2016 code.
While not totally necessary, it is fast and provides a comparison without degrading the original version.
A tidbit on how PID (PI actually) tuning works. 
When we change the offset in the MPU6050 we can get instant results. This allows us to use Proportional and integral of the PID to discover the ideal offsets. Integral is the key to discovering these offsets, Integral uses the error from set-point (set-point is zero), it takes a fraction of this error (error * ki) and adds it to the integral value.   Each reading narrows the error down to the desired offset. The greater the error from set-point, the more we adjust the integral value. The proportional does its part by hiding the noise from the integral math. The derivative is not used because of the noise and because the sensor is stationary.  With the noise removed the integral value lands on a solid offset after just 600 readings. At the end of each set of 100 readings, the integral value is used for the actual offsets and the last proportional reading is ignored due to the fact it reacts to any noise.