Overflow protection and no interrupts
Eliminated the headakes from FIFO buffer overflow and No More need for the interrupt pin and attach interrupts

After an overflow of the FIFO buffer I correct the damage to the corrupted FIFO buffer eliminating the need to reset the fifo buffer. Just get the latest DMP reading when you need it ignore the rest.
One simple function call
GetCurrentFIFOPacket() if you need data onec in a blue moon
and
GetCurrentFIFOPacketTimed() if you want to have the speed of using the interupt pin but don't want to deal with interrupts uses a blink withou delay timer to capture the fifo packet at the moment it is created (auto tunes).

Now you can dely() all you want and still get perfect results!!!
Z