Added a single functrion to handle gathering the latest packet from the DMP generated FIFO Buffer
The Added  dmpGetCurrentFIFOPacket function:
- Handles overflow and all other conditions to retrieve the latest packes in the most efficient way.
- This will work with or without interrupt triggering and
- All blocking code has no affect on the results as the latest packet is still retrieved after any and all delays no matter how long.
- This has been tested over the past several months without any further modifications and should be ready for prime time.

The example code has been modified and now uses this function as it is more efficient and more reliable than the current process.

ToDo Test the function with ESP32 for the ESPWiFi sketch. and make changes in that example which I believe this would seamlessly improve that sketch once tested.
ZHomeSlice